chore: split e2e
