+++ integration/document-preview.spec.ts
@@ -1,24 +1,23 @@
 import { Selector } from "testcafe";
 import { enterPassword, loadConfigFile } from "./helper";
 
-fixture("Document Creator").page`http://localhost:3000`;
+fixture("Document preview").page`http://localhost:3000`;
 
-const Config = "./../src/test/fixtures/sample-local-config.json";
-const DataFile = "./../src/test/fixtures/sample-data-file.json";
+const Config = "./../src/test/fixtures/sample-config-local.json";
+const DataFile = "./../src/test/fixtures/sample-data-file-coo.json";
 const Title = Selector("h1");
-const Button = Selector("button");
-const PasswordField = Selector("[data-testid='password-field']");
 const ProgressBar = Selector("[data-testid='progress-bar']");
+
+const Button = Selector("button");
 const Iframe = Selector("#iframe[title='Decentralised Rendered Certificate']");
 const IframeRoot = Selector("#root");
 
-test("Preview form with data", async (t) => {
-  // upload config and reset config file
+test("should be able to preview form with data", async (t) => {
+  // Upload config file
   await loadConfigFile(Config);
   await t.expect(Title.textContent).contains("Create Document");
 
-  // login to step 1
-  await t.selectText(PasswordField);
+  // Login to step 1
   await enterPassword("password");
   await t.expect(Title.textContent).contains("Choose Document Type to Issue");
   await t.expect(ProgressBar.textContent).contains("Step 1/3");
@@ -27,6 +26,8 @@ test("Preview form with data", async (t) => {
   await t.click(Button.withText("COO"));
   await t.expect(Title.textContent).contains("Fill and Preview Form");
   await t.expect(ProgressBar.textContent).contains("Step 2/3");
+
+  // Upload data file
   await t.setFilesToUpload("input[type=file][data-testid=config-file-drop-zone]", [DataFile]);
 
   // Set preview mode to true
