+++ integration/error-attachment-limit.spec.ts
@@ -1,45 +1,34 @@
 import { Selector } from "testcafe";
 import { enterPassword, loadConfigFile } from "./helper";
 
-fixture("Document Creator").page`http://localhost:3000`;
+fixture("Error attachment limit").page`http://localhost:3000`;
 
-const Config = "./../src/test/fixtures/sample-local-config.json";
+const Config = "./../src/test/fixtures/sample-config-local.json";
 const AttachmentSampleThatIs6Mb = "./../src/test/fixtures/sample-file-6MB.pdf";
 const AttachmentSample = "./../src/test/fixtures/sample.pdf";
 
 const Title = Selector("h1");
 const Button = Selector("button");
 const FormIdField = Selector("#root_iD");
-const ButtonLogin = Selector("[data-testid='login-button']");
-const PasswordField = Selector("[data-testid='password-field']");
-const PasswordFieldMsg = Selector("[data-testid='password-field-msg']");
 const ProgressBar = Selector("[data-testid='progress-bar']");
 const FileSizeError = Selector("[data-testid='file-size-error']");
 const FormAttachmentField = Selector("[data-testid='upload-file-0']");
 
-test("Upload configuration file and test password field to handle correctly", async (t) => {
-  // load config
+test("should show file limit warning when over 6mb", async (t) => {
+  // Upload config file
   await loadConfigFile(Config);
+  await t.expect(Title.textContent).contains("Create Document");
 
-  //try login without password
-  await t.click(ButtonLogin);
-  await t.expect(PasswordFieldMsg.textContent).contains("Invalid password. Please try again.");
-
-  // try login with wrong password
-  await enterPassword("test error");
-  await t.expect(PasswordFieldMsg.textContent).contains("Invalid password. Please try again.");
-
-  // login again with the correct password
-  await t.selectText(PasswordField);
+  // Login to step 1
   await enterPassword("password");
   await t.expect(Title.textContent).contains("Choose Document Type to Issue");
   await t.expect(ProgressBar.textContent).contains("Step 1/3");
 
-  // select a form
+  // Navigate to form
   await t.click(Button.withText("COO"));
   await t.typeText(FormIdField, "COO-ID");
 
-  // upload a file that is greater than 5MB and expect an error
+  // Upload a attachment (over file limit)
   await t.setFilesToUpload("input[data-testid=attachment-file-drop-zone]", [
     AttachmentSampleThatIs6Mb,
   ]);
@@ -47,7 +36,7 @@ test("Upload configuration file and test password field to handle correctly", as
     .expect(FileSizeError.textContent)
     .contains("Error: Total attachment file size exceeds 5MB");
 
-  // upload a proper file and the error is gone
+  // Upload a attachment (below file limit)
   await t.setFilesToUpload("input[data-testid=attachment-file-drop-zone]", [AttachmentSample]);
   await t.expect(FormAttachmentField.textContent).contains("sample.pdf");
 });
