+++ integration/error-configs.spec.ts
@@ -1,27 +1,27 @@
 import { Selector } from "testcafe";
 import { loadConfigFile } from "./helper";
 
-fixture("Document Creator").page`http://localhost:3000`;
+fixture("Error configs").page`http://localhost:3000`;
 
-const Config = "./../src/test/fixtures/sample-local-config.json";
-const ConfigWithError = "./../src/test/fixtures/sample-error-config.json";
-const ConfigErrorFile = "./../src/test/fixtures/sample-empty-error-config.json";
+const Config = "./../src/test/fixtures/sample-config-local.json";
+const ConfigWithError = "./../src/test/fixtures/sample-config-error-walletless.json";
+const ConfigErrorFile = "./../src/test/fixtures/sample-config-error-empty.json";
 
 const Title = Selector("h1");
 const ButtonReset = Selector("[data-testid='reset-button']");
 const ErrorCantReadFile = Selector("[data-testid='error-cannot-read-file']");
 const ConfigError = Selector("[data-testid='config-error']");
 
-test("Upload configuration file with error to test the errors", async (t) => {
-  // upload invalid config file(without wallet)
+test("should show correct error messages on various malformed configs", async (t) => {
+  // Upload config file (without wallet)
   await loadConfigFile(ConfigWithError);
   await t.expect(ConfigError.textContent).contains("Config is malformed");
 
-  // upload invalid file that is not a config file
+  // Upload config file (invalid config file)
   await loadConfigFile(ConfigErrorFile);
   await t.expect(ErrorCantReadFile.textContent).contains("File cannot be read");
 
-  // upload config and reset config file
+  // Upload config file (working config file)
   await loadConfigFile(Config);
   await t.expect(Title.textContent).contains("Create Document");
   await t.expect(Selector("[data-testid='login-title']").textContent).contains("Login");
