+++ src/test/fixtures/sample-config-error-empty.json
