+++ src/test/fixtures/sample-config-error-walletless.json
